Move behavior BattleshipGUI to Player tree
