--- src/main/justacommonguy/battleshipgui/ClientPlayer.java
@@ -1,8 +1,0 @@
-package justacommonguy.battleshipgui;
-
-public class ClientPlayer extends Player{
-
-	public ClientPlayer(String name) {
-		super(name);
-	}
-}
