+++ src/main/justacommonguy/battleshipgui/EnemyPlayer.java
@@ -1,0 +1,11 @@
+package justacommonguy.battleshipgui;
+
+import justacommonguy.battleshipgui.gui.EnemyCell;
+
+public class EnemyPlayer extends Player<EnemyCell> {
+
+	public EnemyPlayer(String name) {
+		super(name, Faction.ENEMY);
+	}
+	
+}
